Updating Azure GO SDK and project folder/module structure
