+++ keyvaultcertdownloader/internal/models/go.mod
@@ -1,0 +1,3 @@
+module models
+
+go 1.18
