--- src/go.mod
@@ -1,13 +1,0 @@
-module github.com/microsoft/keyvaultcertdownloader/src
-
-go 1.13
-
-require (
-	github.com/Azure/azure-sdk-for-go v40.4.0+incompatible
-	github.com/Azure/go-autorest/autorest v0.10.0
-	github.com/Azure/go-autorest/autorest/azure/auth v0.4.2
-	github.com/Azure/go-autorest/autorest/to v0.3.0 // indirect
-	github.com/Azure/go-autorest/autorest/validation v0.2.0 // indirect
-	golang.org/x/crypto v0.0.0-20191206172530-e9b2fee46413
-	software.sslmate.com/src/go-pkcs12 v0.0.0-20190322163127-6e380ad96778
-)
